实现阅读数

+++ src/main/java/com/example/forum/model/User.java
@@ -81,7 +81,7 @@ public class User {
      *
      * @mbg.generated Thu Sep 05 06:46:42 CST 2019
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
